+++ test/driver/measure.cpp
@@ -109,11 +109,11 @@ try {
 
       // Write data.
       const auto begin = chrono::system_clock::now();
-      const auto row_count = data.size();
-      const auto col_count = data.channel_count();
+      const auto row_count = data.row_count();
+      const auto col_count = data.column_count();
       for (std::size_t row{}; row < row_count; ++row) {
         for (std::size_t col{}; col < col_count; ++col)
-          log_file << data[col][row] << " ";
+          log_file << data.value(col, row) << " ";
         log_file << "\n";
       }
       if (eco)